requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
dnspython==2.3.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.21.6
pandas==1.3.5
PyJWT==2.6.0
pymongo==4.3.3
pyreadline==2.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.11.0
